requirements.txt
flask==3.0.2
pillow==10.2.0
qrcode==7.4.2
python-dotenv==1.0.0
